+++ src/apps/LavaMail/controllers/ctrlCompose.js
@@ -90,6 +90,15 @@ module.exports = ($rootScope, $scope, $stateParams, $translate,
 			let people = [...contacts.people.values()];
 			let map = new Map();
 
+			const addHiddenEmails = (checkEmail) => {
+				people.reduce((a, c) => {
+					if (c.hiddenEmail && c.hiddenEmail.email && checkEmail(c.hiddenEmail))
+						a.set(c.hiddenEmail.email, c.hiddenEmail);
+
+					return a;
+				}, map);
+			};
+
 			const addEmails = (checkEmail) => {
 				people.reduce((a, c) => {
 					c.privateEmails.forEach(e => {
@@ -109,6 +118,9 @@ module.exports = ($rootScope, $scope, $stateParams, $translate,
 			addEmails(e => e.isSecured() && !e.isStar);
 			addEmails(e => !e.isSecured() && e.isStar);
 			addEmails(e => !e.isSecured() && !e.isStar);
+
+			addHiddenEmails(e => e.isSecured());
+			addHiddenEmails(e => !e.isSecured());
 
 			if (toEmailContact)
 				map.set(toEmailContact.email, toEmailContact);
@@ -344,9 +356,9 @@ module.exports = ($rootScope, $scope, $stateParams, $translate,
 			yield manifest.getDestinationEmails()
 				.filter(email => !contacts.getContactByEmail(email))
 				.map(email => {
-					return contacts.createContact(new Contact({
-						privateEmails: [hiddenContacts[email]]
-					}));
+					let contact = new Contact({name: '$hidden'});
+					contact.hiddenEmail = hiddenContacts[email];
+					return contacts.createContact(contact);
 				});
 
 			manifest = null;
@@ -427,7 +439,14 @@ module.exports = ($rootScope, $scope, $stateParams, $translate,
 		if (contacts.getContactByEmail(email))
 			return null;
 
-		newHiddenContact = ContactEmail.newHiddenEmail(email, email, newTag);
+		newHiddenContact = new ContactEmail(null, {
+			isTag: true,
+			tag: newTag,
+			name,
+			email,
+			isNew: true
+		}, 'hidden');
+
 		newHiddenContact.loadKey();
 
 		hiddenContacts[newHiddenContact.email] = newHiddenContact;