+++ src/apps/LavaMail/services/contacts.js
@@ -36,6 +36,18 @@ module.exports = function($q, $rootScope, co, user, crypto, utils, LavaboomAPI, 
 		}, new Map());
 	});
 
+	const removedReplacedHiddenEmails = (contact) => co(function *(){
+		const emails = utils.uniq([...contact.privateEmails.map(e => e.email), ...contact.businessEmails.map(e => e.email)]);
+		for(let e of emails) {
+			const c = self.getContactByEmail(e);
+			if (c.isHidden()) {
+				yield self.deleteContact(c.id);
+
+				break;
+			}
+		}
+	});
+
 	this.createContact = (contact) => co(function *() {
 		let envelope = yield Contact.toEnvelope(contact);
 		let r = yield LavaboomAPI.contacts.create(envelope);
@@ -52,6 +64,8 @@ module.exports = function($q, $rootScope, co, user, crypto, utils, LavaboomAPI, 
 
 		contact.id = r.body.contact.id;
 		self.people.set(contact.id, contact);
+		
+		yield removedReplacedHiddenEmails(contact);
 
 		$rootScope.$broadcast('contacts-changed');
 
@@ -61,6 +75,8 @@ module.exports = function($q, $rootScope, co, user, crypto, utils, LavaboomAPI, 
 	this.updateContact = (contact) => co(function *() {
 		let envelope = yield Contact.toEnvelope(contact);
 		let r = yield LavaboomAPI.contacts.update(contact.id, envelope);
+
+		yield removedReplacedHiddenEmails(contact);
 
 		$rootScope.$broadcast('contacts-changed');